requirements.txt
flask>=3.0
flask-sock>=0.7
requests>=2.31
urllib3>=2.0
pystray>=0.19
Pillow>=10.0